requirements.txt
PyQt5==5.15.7
python-vlc==3.0.16120
